Added WebRTC cmake scripts Added cmake macros for enhanded library finding to dry up build scripts Introduced new Logger macros and updated most files TURN server now supports external IP option Symple server now uses string based rather than object based addresses for message size reduction Added WebRTC peerconnection sample which integrates with the LibSourcey media layer Added an IPC message queue to the base library

+++ src/http/include/scy/http/websocket.h
@@ -167,7 +167,7 @@ public:
 	virtual int writeFrame(const char* buffer, int length, int flags, BitWriter& frame);
 		// Writes a WebSocket protocol frame from the given data.
 		
-	virtual int readFrame(BitReader& frame, char*& payload); //Buffer& buffer, const char* buffer, int length, 
+	virtual UInt64 readFrame(BitReader& frame, char*& payload); //Buffer& buffer, const char* buffer, int length, 
 		// Reads a single WebSocket frame from the given buffer (frame).
 		//
 		// The actual payload length is returned, and the beginning of the
